### 2025-12-15 #193 @editorialbot
[PRE REVIEW]: Bringing Julia to Health: A Concise Overview of JuliaHealth
**Submitting author:** @TheCedarPrince (<a href="http://orcid.org/0000-0003-4531-1614">Jacob Zelko</a>)
**Repository:** https://github.com/JuliaHealth/JuliaHealthEcosystemPaper
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
**Managing EiC:** Dimitris Floros

Thanks for submitting your paper to JCON @TheCedarPrince. **Currently, there isn't a JCON editor assigned** to your paper.
