### 2025-12-24 #193 @pitsianis
> Interested in reviewing but zero experience here. What is the deadline?

There is a first time for everything.
The whole process is transparent and typified. Follow the instructions, and all will go well.
We are also here to advise and assist.

There are no hard deadlines, but it will be good to wrap it up in a month if all parties are responsive.

Shall I add you as a reviewer?
