### 2026-01-05 #193 @TheCedarPrince
Hey @pitsianis , looking at the list of folks from the JuliaCon Proceedings reviewer list, let me ping a few folks:

- @apraga
- @ryanguo13
- @agchesebro
- @kellertuer
- @Tortar

Folks, thanks for being potential reviewers. For context, this 2 page paper is an extended abstract paper describing a concise and consolidated summary of the JuliaHealth ecosystem based on discussions had at the first annual JuliaHealth mini-Symposium at JuliaCon 2025.

If you would be willing to review, please comment below. Otherwise, apologies for the ping.

Thank you!

~ tcp 🌳
JuliaHealth Lead
