### 2026-01-05 #193 @agchesebro
Happy to review, but would likely get to it next weekend. Let me know if you prefer faster turnaround!
