### 2026-01-05 #194 @editorialbot
[REVIEW]: Bringing Julia to Health: A Concise Overview of JuliaHealth
**Submitting author:** @TheCedarPrince (<a href="http://orcid.org/0000-0003-4531-1614">Jacob Zelko</a>)
**Repository:** https://github.com/JuliaHealth/JuliaHealthEcosystemPaper
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.0
