### 2026-01-25 #194 @apraga
Here's my review: well written and interesting to see such efforts. It's a paper worth publishing. Minor remarks
1. There is no statement of need. It seems the community was created organically but was there a need for such an initiative ? Are current tools not open-source or is a global framwork needed ? Also, why Julia ? The functionality is quite clear, but not what issue it aims to taclke.
2. Terms and abrevations should be defined somewhere : the DICOM and OMOP format and the OHDSI.

@pitsianis the checklist is not suitable for this proposal. There is no single software, but several packages in a shared organization. Reviewing each package seems out-of-scope and time-consuming. Also, I think the github link should point to https://github.com/JuliaHealth.

Thanks,
