### 2026-01-26 #194 @agchesebro
Sorry for the delay - finally got to this! Agree this is a good overview of the ecosystem and worth publishing. A few minor thoughts for updates:

1. I think COVID-19 is hyphenated by convention (currently COVID19 in the manuscript).
2. I’d rephrase the part in 2.1.1 from “measuring neuron currents” to “sensitively measuring oscillating magnetic currents in the brain.” This is a minor gripe but up my alley so I’m pushing for a more precise phrase of what it’s actually doing (as an aside I’ve been loving the KomaMRI work and it’s been fun to play with for my research :).
3. In 2.2 “profiels” —> profiles
4. It’d be nice to have a compiled PDF of the final paper in the repo that this points to just so readers don’t have to compile it on their own.
5. Agree with @apraga that the repo should link to the main JuliaHealth organization at some point just to help guide readers.
