### 2026-01-29 #194 @TheCedarPrince
Hi @apraga and @agchesebro:

Thank you very much for your reviews and critiques!

We have addressed all comments you all raised in the following way for this paper:

## Syntax, Typos and Abbreviations

> Terms and abrevations should be defined somewhere : the DICOM and OMOP format and the OHDSI.

These acronyms have been explained and defined inline within the paper.

> I think COVID-19 is hyphenated by convention (currently COVID19 in the manuscript).

This convention has been updated across the paper.

> In 2.2 “profiels” —> profiles

Fixed!

## Structural Concerns

> There is no statement of need. It seems the community was created organically but was there a need for such an initiative ? Are current tools not open-source or is a global framwork needed ? Also, why Julia ? The functionality is quite clear, but not what issue it aims to taclke.

This comment has been addressed as a new, concise section titled, "Statement of Need" after the "Introduction". The section addresses all concerns about need, why Julia, and current tooling; let us know if that is satisfactory to your concerns or not @apraga

> Agree with @apraga that the repo should link to the main JuliaHealth organization at some point just to help guide readers.

We have included this reference multiple times throughout the paper to make the link clear.

## Misc

> It’d be nice to have a compiled PDF of the final paper in the repo that this points to just so readers don’t have to compile it on their own.

We have included the compiled PDF in the repo

---

Much thanks to @apraga and @agchesebro for their reviews to improve this paper and great appreciation to @kosuri-indu and @cncastillo for spearheading revisions!

Please do not hesitate to let us know of any remaining concerns.

~ tcp
