### 2026-02-09 #194 @luraess
Thank you for finalising the revision and implementing the reviewers' comments @TheCedarPrince .

Before we can publish the proceeding paper, it would be ideal if you could address the following steps which would enhance the reach and improve the rendering of the paper:
- Modify the hyperlinks from boxes to coloured items
- If possible make the ticks for the list items sorter and detached from the text
- Refer to GitHub packages and orgs with hyperlinks when possible and if you have a Zenodo (or similar release) for citation that could be a plus.
- Providing a verbose link to https://juliahealth.org may be a nice addition (possibly adding the logo as small figure to the paper?)
- We may want an archive (Zenodo release or similar) for the main assets related to your paper. Please provide this as well.

Once the above steps are addressed, we could finalise. Thanks for your help and contribution!
