### 2026-03-05 b7e5e17
Finished editor suggestions
