+++ paper.tex
@@ -1,6 +1,13 @@
+% At the very top, before \documentclass
+\PassOptionsToPackage{colorlinks=true, linkcolor=blue, urlcolor=cyan, citecolor=green}{hyperref}
+
 % JuliaCon proceedings template
 \documentclass{juliacon}
 \setcounter{page}{1}
+
+\usepackage{enumitem}
+\usepackage{wrapfig}
+
 
 \begin{document}
 
@@ -32,8 +39,13 @@ We conclude by outlining future development goals and directions for community g
 
 \section{Introduction}
 
+\begin{wrapfigure}{l}{0.05\textwidth}
+  \centering
+  \includegraphics[width=0.075\textwidth]{juliahealth_logo_small.png}
+\end{wrapfigure}
+
 Founded in 2020, JuliaHealth was created to support the development and adoption of high-performance computing methods from the Julia ecosystem for biomedical and health research.
-The community's packages are developed under the \href{https://github.com/JuliaHealth}{JuliaHealth GitHub organization}.
+The community's packages are developed under the \href{https://github.com/JuliaHealth}{JuliaHealth GitHub organization} with additional community details at \href{https://juliahealth.org}{juliahealth.org}.
 Initially, JuliaHealth served as a coordination point for Julia community efforts related to the COVID-19 pandemic.
 Since then, JuliaHealth has continued to grow, bringing together dozens of active contributors across its GitHub repositories and a broader user community across Julia communication channels, and supporting research and professional activities across multiple venues \cite{zelko2025juliahealth, Zelko2023Julia}.
 As the community has expanded, several sub-communities have organically emerged within JuliaHealth to address different aspects of health research, each with distinct computational needs and levels of maturity.
@@ -45,11 +57,9 @@ The purpose of this paper is to describe these emerging ecosystems, their curren
 
 Biomedical and health research increasingly relies on computational workflows that must be open, reproducible, and efficient.
 However, the software ecosystem supporting these workflows remains fragmented across closed-source platforms and multiple programming languages.
-Researchers commonly combine tools written in MATLAB, Python, R, or C/C++, resulting in workflows that are difficult to integrate, reproduce, and maintain, particularly when performance-critical components are involved.
-
+Researchers commonly combine tools written in MATLAB, Python, R, or C/C++, resulting in workflows that are difficult to integrate, reproduce, and maintain, particularly when performance-critical components are involved. 
 Julia provides a suitable technical basis for addressing these challenges because it supports high-level scientific programming with performance approaching low-level languages, enabling researchers to prototype and deploy within a single language.
-When external tools are needed, Julia can also reuse established biomedical software through interoperability packages such as PythonCall.jl and RCall.jl.
-
+When external tools are needed, Julia can also reuse established biomedical software through interoperability packages such as PythonCall.jl \cite{PythonCall.jl} and RCall.jl \cite{randy_lai_2026_18422923}.
 JuliaHealth addresses the need for a community-maintained, open-source framework that spans multiple packages and can be composed into end-to-end workflows (e.g., simulation to analysis, or cohort definition to study execution) while remaining transparent and reproducible.
 By coordinating development within a shared organization, JuliaHealth lowers the cost of discovering, combining, and maintaining interoperable tools for health research.
 
@@ -68,15 +78,15 @@ The ecosystem spans physics simulations, data acquisition, and reconstruction to
 \subsubsection{Image Acquisition and Simulation}  
 
[…]
