+++ ref.bib
@@ -132,3 +132,107 @@
   journal={arXiv preprint arXiv:2408.15655},
   year={2024}
 }
+
+@misc{PythonCall.jl,
+  author = {Rowley, Christopher},
+  title = {PythonCall.jl: Python and Julia in harmony},
+  year  = {2022},
+  url   = {https://github.com/JuliaPy/PythonCall.jl},
+}
+
+@software{randy_lai_2026_18422923,
+  author       = {Randy Lai and
+                  Simon Byrne and
+                  Douglas Bates and
+                  Phillip Alday and
+                  Viral B. Shah and
+                  Milan Bouchet-Valat and
+                  Frankie Robertson and
+                  Dilum Aluthge and
+                  Dave Kleinschmidt and
+                  Changcheng Li and
+                  Ranjan Anantharaman and
+                  lbilli and
+                  Antoine Baldassari and
+                  Steven G. Johnson and
+                  Páll Haraldsson and
+                  Alex Arslan and
+                  Robert Feldt and
+                  Pietro Monticone and
+                  Michael Hatherly and
+                  Lukas Elmiger and
+                  Iain Dunning and
+                  Gerhard Aigner and
+                  Diego Javier Zea and
+                  Avik Sengupta and
+                  Kevin Squire and
+                  Mark Kittisopikul and
+                  Morten Piibeleht and
+                  Patrick Kofod Mogensen},
+  title        = {JuliaInterop/RCall.jl: v0.14.12},
+  month        = jan,
+  year         = 2026,
+  publisher    = {Zenodo},
+  version      = {v0.14.12},
+  doi          = {10.5281/zenodo.18422923},
+  url          = {https://doi.org/10.5281/zenodo.18422923},
+  swhid        = {swh:1:dir:83356e1579dffd910ef14676c29cad752e6ada56
+                   ;origin=https://doi.org/10.5281/zenodo.12771368;vi
+                   sit=swh:1:snp:cc3130ff8177371482bdf68ca3b8b943bf52
+                   9330;anchor=swh:1:rel:665eb96fad5a594e5bc4cf545f38
+                   b8b4271fe04b;path=JuliaInterop-RCall.jl-7b2b73e
+                  },
+}
+
+@Article{Mitura2021,
+  author   = {Mitura, Jakub and Chrapko, Beata E.},
+  journal  = {Zeszyty Naukowe WWSI},
+  title    = {{3D Medical Segmentation Visualization in Julia with MedEye3d}},
+  year     = {2021},
+  number   = {25},
+  pages    = {57--67},
+  volume   = {15},
+  doi      = {10.26348/znwwsi.25.57},
+  keywords = {OpenGl, Computer Tomagraphy, PET/CT, medical image annotation, medical image visualization},
+}
+
+@software{Zelko_OMOPCDMCohortCreator_0_2_0_2022,
+author = {Zelko, Jacob and Chinta, Varshini},
+doi = {10.5281/zenodo.7052105},
+month = {10},
+title = {{OMOPCDMCohortCreator 0.2.0}},
+version = {0.2.0},
+year = {2022}
+}
+
+
+@misc{OMOPCDMPathways.jl,
+	author  = {Landge, Jay Sanjay and S. Zelko, Jacob and contributors},
+	title   = {OMOPCDMPathways.jl},
+	url     = {https://github.com/JuliaHealth/OMOPCDMPathways.jl},
+	version = {v1.0.0-DEV},
+	year    = {2024},
+	month   = {5}
+}
+
+
+@article{mannes2024derivation,
[…]
