### 2026-03-05 #194 @TheCedarPrince
Hi @luraess , thanks for the comments -- these have all been addressed! Here is how:

> Modify the hyperlinks from boxes to coloured items

Done -- changed to blue hyperlinks.

> If possible make the ticks for the list items sorter and detached from the text

Done, cleaned this and made it more compact.

> Refer to GitHub packages and orgs with hyperlinks when possible and if you have a Zenodo (or similar release) for citation that could be a plus.

Updated.

> Providing a verbose link to [juliahealth.org](https://juliahealth.org/) may be a nice addition (possibly adding the logo as small figure to the paper?)

Added both the link explicitly in the introduction and a small logo figure inline at the start of the paper.

> We may want an archive (Zenodo release or similar) for the main assets related to your paper. Please provide this as well.

Here is the Zenodo link: https://zenodo.org/records/18879155

---

Let me know if anything else is needed for finalization! Thank you!
