### 2026-03-07 #194 @editorialbot
:wave: @JuliaCon/jcon-eics, this paper is ready to be accepted and published.
