### 2026-03-07 #194 @luraess
@TheCedarPrince all looks good from my side now - thanks for addressing the remaining suggestions.

Can you please check that the final proof is OK for you and then we could be fine for publishing.

One thing you may want to check in particular are the missing DOIs for some of your refs. If you can find them that'd be awesome. If non existant, fine as well.

Thanks
