### 2026-03-07 ceca48a
Fixed up DOIs and added biomedical signal processing citation
