+++ paper.tex
@@ -105,7 +105,7 @@ Beyond these two ecosystems, many parts of JuliaHealth are actively under develo
 For example, for simulation of biological and physiological processes, Thunderbolt.jl simulates cardiac electrophysiology and mechanics, and BloodFlowTrixi.jl can be used to explore hemodynamics. \cite{mannes2024derivation, mannes2024derivationPhD, ersoy2024new}
 For interacting with third-party bioinformatics platforms, several API wrappers and specialized tools exist to bring relevant data or information from biomedical literature and chemical databases into one's workflow.
 These tools support programmatic retrieval and analysis via packages such as \href{https://github.com/JuliaHealth/PubMedMiner.jl}{PubMedMiner.jl} for accessing literature from PubMed/MEDLINE and \href{https://github.com/JuliaHealth/PubChemCrawler.jl}{PubChemCrawler.jl} for retrieving chemical structures and properties from PubChem.
-As JuliaHealth continues to grow, these efforts in simulation modeling and integration with bioinformatics resources will help expand JuliaHealth into new ecosystems and further stages of development.
+As JuliaHealth continues to grow, these efforts in simulation modeling and integration with bioinformatics resources will help expand JuliaHealth into new ecosystems and further stages of development. \cite{datseris2024physiological}
 
 \section{Discussion}
 
