+++ ref.bib
@@ -219,6 +219,8 @@ year = {2022}
 @article{mannes2024derivation,
   title={ON THE DERIVATION OF A NEW ONE-DIMENSIONAL MODEL FOR BLOOD FLOWS AND ITS NUMERICAL APPROXIMATION},
   author={Mannes, Yolhan and Ersoy, Mehmet and Eker, Omer Faruk},
+  doi = {10.4236/jamp.2025.1310198},
+  journal = {Journal of Applied Mathematics and Physics},
   year={2024}
 }
 
@@ -226,6 +228,9 @@ year = {2022}
 @article{ersoy2024new,
   title={A NEW TWO-DIMENSIONAL BLOOD FLOW MODEL AND ITS RKDG APPROXIMATION},
   author={Ersoy, Mehmet and Mannes, Yolhan and Eker, Omer},
+  doi = {10.4236/jamp.2025.1311220},
+  journal = {Journal of Applied Mathematics
+and Physics},
   year={2024}
 }
 
@@ -235,4 +240,16 @@ year = {2022}
   author={Mannes, Yolhan},
   year={2024},
   school={Universit{\'e} de Toulon}
+}
+
+
+@article{datseris2024physiological,
+  title={Physiological signal analysis and open science using the Julia language and associated software},
+  author={Datseris, George and Zelko, Jacob S},
+  journal={Frontiers in Network Physiology},
+  volume={4},
+  pages={1478280},
+  year={2024},
+  publisher={Frontiers Media SA},
+  doi = {10.3389/fnetp.2024.1478280}
 }
