### 2026-03-07 #194 @TheCedarPrince
Hey @luraess , I updated the paper with DOI references and after you rebuild the final proof, this looks perfect to me!

Note: I did try to find the missing DOIs and I was able to find most of them but a few of them did not have associated DOIs.

Thanks!
